- fixed some warnings - SAD SSE implementation for u8
git-svn-id: https://svnknoll.informatik.tu-muenchen.de/kloses/visionTools@923 f7bcf5df-4c73-0410-a0d8-b61c104701fe

+++ cvt/util/SIMD.cpp
@@ -1076,7 +1076,7 @@ namespace cvt {
 		return sad;
 	}
 	
-	float SIMD::SAD( uint8_t const* src1, uint8_t const* src2, const size_t n ) const
+	size_t SIMD::SAD( uint8_t const* src1, uint8_t const* src2, const size_t n ) const
 	{
 		size_t i = n >> 2;
